added Batch rendering & animation (WIP)

+++ kgEngine-2.0/TileMapGame/Transformation.cpp
@@ -129,7 +129,7 @@ namespace kg
 			setPosition(
 				Vector2i( atoi( information.at( 0 ).c_str() ),
 				atoi( information.at( 1 ).c_str() ) ) );
-			setRotation( atof( information.at( 2 ).c_str() ) );
+			setRotation( static_cast<float>(atof( information.at( 2 ).c_str() )) );
 			setSize( Vector2i(
 				atoi( information.at( 3 ).c_str() ),
 				atoi( information.at( 4 ).c_str() ) ) );
